+++ afanasy/trunk/src/libafanasy/taskexec.h
@@ -130,9 +130,9 @@ public:
       UPLAST
    };
 
-   inline bool     addListenAddress( const Address * address) { return listen_addresses->addAddress( address);     }
-   inline bool  removeListenAddress( const Address * address) { return listen_addresses->removeAddress( address);  }
-   inline int   getListenAddressesNum() const { return listen_addresses->getAddressesNum();              }
+   inline bool         addListenAddress( const Address * address) { return listen_addresses->addAddress( address);     }
+   inline bool      removeListenAddress( const Address * address) { return listen_addresses->removeAddress( address);  }
+   inline size_t getListenAddressesNum() const { return listen_addresses->getAddressesNum();              }
    inline const std::list<Address*> * getListenAddresses() const { return listen_addresses->getAddresses();       }
 
    inline void setName(      const std::string & str) { name      = str;}   ///< Set task name.
